% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{censoringData}
\alias{censoringData}
\title{Adding rows with censored data points}
\usage{
censoringData(
  dat,
  nestVars,
  timeVar,
  eventVar = "event",
  catVar = NULL,
  TimeGap = 1,
  censoring = "right",
  verbose = T
)
}
\arguments{
\item{dat}{data.frame containing the variables to be lagged}

\item{nestVars}{name(s) of the columns indicating how the data is nested
(e.g., ID variable). Currently up to two \code{nestVars} are possible.}

\item{timeVar}{name of the column with the time variable.}

\item{eventVar}{name of the column indicating if an event (value = 1) happened or not/censored (0)}

\item{catVar}{name of the event (for coxph or frailty) or event-type (for multi-state)
column where the indication of the censoring is stored.}

\item{censoring}{character string \code{"right"} or \code{"left"} indicating if right- or left-
censored data row should be added}

\item{timeGap}{time to be added (for right-censoring) or removed (for left-censoring) form \code{timeVar}}
}
\value{
data.frame with extra rows for censored data.
}
\description{
Creates extra rows for indicating left- or right-censored data.
}
\examples{
 dat <- data.frame(ID = c(rep(1,10), rep(2,10)),
                  day = rep(c(rep(1,5), rep(2,5)),2),
                  event = 1,
                  a = runif(20, max = 10),
                  t = rlnorm(20,1,.4),
                  cat = sample(c("X","Y","Z"), 20, replace = T))
censoringData(dat, nestVars = c("ID"), timeVar = "t")
censoringData(dat, nestVars = c("ID","day"), timeVar = "t")
censoringData(dat, nestVars = c("ID"), timeVar = "t", censoring = "left",TimeGap = 1

}
