% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{defineMorningMeasure}
\alias{defineMorningMeasure}
\title{MAIN TITLE}
\usage{
defineMorningMeasure(dat, dayVar, nestVars)
}
\arguments{
\item{dat}{a data.frame object}

\item{dayVar}{}

\item{nestVars}{either a character object or a vector of character objects with the column names
in the dat object, indicating the nesting of the data (e.g., participant ID). Up to three nesting layers can be defined.}
}
\value{

}
\description{
initial description
}
\examples{

}
