% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{getAbsTime}
\alias{getAbsTime}
\title{MAIN TITLE}
\usage{
getAbsTime(
  dat,
  nestVars = "id",
  timeVar = "time",
  origin = Sys.time(),
  verbose = T,
  ...
)
}
\arguments{
\item{dat}{a data.frame object}

\item{nestVars}{either a character object or a vector of character objects with the column names
in the dat object, indicating the nesting of the data (e.g., participant ID). Up to three nesting layers can be defined.}

\item{timeVar}{name of the column in the dat object indicating the time of observation in POSIX format.}

\item{verbose}{should processing information be printed during the functions run time. Default is FALSE.}
}
\value{

}
\description{
initial description
}
\examples{

}
