% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getTransitionMatrix}
\alias{getTransitionMatrix}
\title{MAIN TITLE}
\usage{
getTransitionMatrix(dat, catVar, type = "sum", categories = NULL)
}
\value{

}
\description{
initial description
}
\examples{

}
