% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{insertAloneTime}
\alias{insertAloneTime}
\title{creates additional rows for new categories}
\usage{
insertAloneTime(
  dat,
  nestVars = NULL,
  timeVar = "date",
  catVar = "alter",
  insertNA = NULL,
  timeLag = 1
)
}
\arguments{
\item{dat}{data.frame containing the relevant variables}

\item{nestVars}{name(s) of the columns indicating how the data is nested
(e.g., ID variable). Currently up to two \code{nestVars} are possible.}

\item{timeVar}{column name indicating the time variable}

\item{catVar}{column name indicating the category variable}

\item{insertNA}{A vector with the variable names that should be filled 
with NAs for the "not-alone" state}

\item{timeLag}{A positive number or vector with positive numbers indicating 
the time in the "not-alone" state. If a vector is provided random samples of the vector will be taken.}
}
\value{
data.frame with addtional rows
}
\description{
function to add rows for new types of observations  (e.g., Alone).
}
\examples{
dat <- data.frame(ID = c(rep(1,10), rep(2,10)),
                 day = rep(c(rep(1,5), rep(2,5)),2),
                 a = runif(20, max = 10),
                 time = Sys.time()+1:20,
                 cat = sample(c("X","Y","Z"), 20, replace = T))

insertAloneTime(dat, nestVars = c("ID"), timeVar = "time", catVar = "cat")
}
