% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{insertTimeBetween}
\alias{insertTimeBetween}
\title{inserting time between observations}
\usage{
insertTimeBetween(
  dat,
  startVar = "start",
  endVar = "end",
  nestVars = NULL,
  catVar = "int",
  catName = "alone"
)
}
\value{

}
\description{
this function inserts additional rows filling the time between (unobserved) events
}
\examples{

}
