% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{lagVars}
\alias{lagVars}
\title{Lagging variables}
\usage{
lagVars(
  dat,
  lags = 1,
  vars = vars,
  diffvars = diffvars,
  unit = "auto",
  verbose = F,
  ...
)
}
\arguments{
\item{dat}{data.frame containing the variables to be lagged}

\item{lags}{vector defining the \code{n}th lag. Positive values (e.g., 1:2) indicatw 
the values from the previous rows (e.g., the values from the previous (lag = 1) observation) 
should be taken. 
Negative values indicate that the values from the following rows should be taken.}

\item{vars}{names of the columns in the data.frame that should be lagged}

\item{diffvars}{names of columns in the data.frame of which the difference to the current row should be taken
(this is particularly helpful for dealing with time variables and time differences)}

\item{unit}{character value, if time variables are being lagged and the difference is taken, which time unit should be used
to calculate the difference}

\item{verbose}{boolean indicating about the state of the process be displayed}
}
\value{
data.frame with new columns containing the lagged variables
}
\description{
Creates extra column(s) with lagged variables according to the order of the data frame.
}
\examples{
dat <- data.frame(a = runif(10, max = 10),
b = Sys.Date()+1:10)
lagVars(dat, lags = -1:2, vars = c("a","b"), 
diffvars = "b", unit = "hours")

}
\seealso{
\code{\link{lagVarsNested}}
}
