% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{meanCenteringNested}
\alias{meanCenteringNested}
\title{MAIN TITLE}
\usage{
meanCenteringNested(dat = dat, vars, nestVars, verbose = T, na.rm = T)
}
\arguments{
\item{dat}{a data.frame object}

\item{vars}{a vector of variables names, corresponding to columns in the dat object
for which the computation should be carried out}

\item{nestVars}{either a character object or a vector of character objects with the column names
in the dat object, indicating the nesting of the data (e.g., participant ID). Up to three nesting layers can be defined.}

\item{verbose}{should processing information be printed during the functions run time. Default is FALSE.}
}
\value{

}
\description{
initial description
}
\examples{

}
