% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{nestedSurvivalFunction}
\alias{nestedSurvivalFunction}
\title{MAIN TITLE}
\usage{
nestedSurvivalFunction(dat, nestVar, timeVar, eventVar, plot = T, verbose = F)
}
\value{

}
\description{
initial description
}
\examples{

}
