% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.coxme}
\alias{plot.coxme}
\title{MAIN TITLE}
\usage{
\method{plot}{coxme}(fit, fontsize = 12, labels = NULL, order = NULL, ...)
}
\arguments{
\item{labels}{Vector with the labels to be displayed. Provided in the original order (i.e., before ordering)}
}
\value{

}
\description{
initial description
}
\examples{

}
