% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTransitionNetwork}
\alias{plotTransitionNetwork}
\title{MAIN TITLE}
\usage{
plotTransitionNetwork(
  dat,
  catVar,
  type = "sum",
  categories = NULL,
  returnMat = T,
  title = NULL
)
}
\value{

}
\description{
initial description
}
\examples{

}
