% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat2.R
\docType{data}
\name{simdat2}
\alias{simdat2}
\title{Simulated data}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(simdat2)
}
\description{
Simulated survival data for illustration purposes
}
\examples{
data(simdat2)
}
\keyword{datasets}
