% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{lagVars}
\alias{lagVars}
\title{Illustration of crayon colors}
\usage{
lagVars(
  dat,
  lags = lags,
  vars = vars,
  diffvars = diffvars,
  unit = "auto",
  verbose = F,
  ...
)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin parameters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
plot_crayons()

}
